// src/types/costoEnvio.types.ts
export type Provincia =
  | 'BUENOS_AIRES'
  | 'CABA'
  | 'CATAMARCA'
  | 'CHACO'
  | 'CHUBUT'
  | 'CORDOBA'
  | 'CORRIENTES'
  | 'ENTRE_RIOS'
  | 'FORMOSA'
  | 'JUJUY'
  | 'LA_PAMPA'
  | 'LA_RIOJA'
  | 'MENDOZA'
  | 'MISIONES'
  | 'NEUQUEN'
  | 'RIO_NEGRO'
  | 'SALTA'
  | 'SAN_JUAN'
  | 'SAN_LUIS'
  | 'SANTA_CRUZ'
  | 'SANTA_FE'
  | 'SANTIAGO_DEL_ESTERO'
  | 'TIERRA_DEL_FUEGO'
  | 'TUCUMAN';

export interface CostoEnvioDTO {
  id: number | null;
  provincia: Provincia;
  provinciaNombre: string;
  precio: number;
  updatedAt: string | null;
}